+++ src/pysiebanxico/__init__.py
@@ -10,6 +10,7 @@ from .exceptions import (
     SeriesNotFoundError,
 )
 from .models import Observation, SeriesData, SeriesMetadata
+from .utilities import index_by_id, metadata_to_records, to_records
 
 __version__ = "0.1.0.dev0"
 
@@ -24,4 +25,7 @@ __all__ = [
     "SeriesData",
     "SeriesMetadata",
     "SeriesNotFoundError",
+    "index_by_id",
+    "metadata_to_records",
+    "to_records",
 ]